gateway-api: Move Gateway Reconcile Watch functions into module
This commit moves all of the gatewayReconciler Watch functions
away from being methods on that type to functions in a new
`watchhandlers` package.

This substantially reduces the size of the `gateway.go`
file, and is part one of a series of changes aimed at adding more
resources easier.

It also moves some testing fixtures and helpers into a new `testhelpers` module,
as those fixtures and helper functions now need to be used by
multiple tests in multiple modules.
