+++ operator/pkg/gateway-api/helpers/testhelpers/doc.go
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+// testhelpers contains functions and fixtures that are helpful for testing multiple
+// parts of the Gateway API code.
+package testhelpers
