+++ operator/pkg/gateway-api/testdata/gateway/tlsroute-mixed-protocol-listeners/input/tlsroute-mixed-protocol-listeners.yaml
@@ -32,7 +32,7 @@ spec:
       tls:
         mode: Passthrough
 ---
-apiVersion: gateway.networking.k8s.io/v1alpha2
+apiVersion: gateway.networking.k8s.io/v1
 kind: TLSRoute
 metadata:
   name: tlsroute-mixed
