+++ operator/pkg/gateway-api/testdata/gateway/tlsroute-mixed-protocol-listeners/output/tlsroute-tlsroute-mixed.yaml
@@ -1,4 +1,4 @@
-apiVersion: gateway.networking.k8s.io/v1alpha2
+apiVersion: gateway.networking.k8s.io/v1
 kind: TLSRoute
 metadata:
   name: tlsroute-mixed
