+++ operator/pkg/gateway-api/watch-handlers/referencegrant.go
@@ -1,0 +1,50 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+package watchhandlers
+
+import (
+	"context"
+	"log/slog"
+
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
+
+	"github.com/cilium/cilium/pkg/logging/logfields"
+)
+
+// TODO(youngnick): Fix this so it doesn't just enqueue all Gateways whenever a ReferenceGrant
+// is updated.
+
+func EnqueueRequestForReferenceGrant(c client.Client, logger *slog.Logger) handler.EventHandler {
+	return handler.EnqueueRequestsFromMapFunc(enqueueAll(c, logger))
+}
+
+func enqueueAll(c client.Client, logger *slog.Logger) handler.MapFunc {
+	return func(ctx context.Context, o client.Object) []reconcile.Request {
+		scopedLog := logger.With(
+			logfields.Resource, client.ObjectKeyFromObject(o),
+		)
+		list := &gatewayv1.GatewayList{}
+
+		if err := c.List(ctx, list, &client.ListOptions{}); err != nil {
+			scopedLog.ErrorContext(ctx, "Failed to list Gateway", logfields.Error, err)
+			return []reconcile.Request{}
+		}
+
+		requests := make([]reconcile.Request, 0, len(list.Items))
+		for _, item := range list.Items {
+			gw := client.ObjectKey{
+				Namespace: item.GetNamespace(),
+				Name:      item.GetName(),
+			}
+			requests = append(requests, reconcile.Request{
+				NamespacedName: gw,
+			})
+			scopedLog.InfoContext(ctx, "Enqueued Gateway for resource", logfields.Gateway, gw)
+		}
+		return requests
+	}
+}
