+++ operator/pkg/gateway-api/watch-handlers/secrets.go
@@ -1,0 +1,33 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Authors of Cilium
+
+package watchhandlers
+
+import (
+	"context"
+	"log/slog"
+
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+
+	"github.com/cilium/cilium/operator/pkg/gateway-api/helpers"
+)
+
+// EnqueueRequestForTLSSecret returns an event handler for any changes with TLS secrets
+func EnqueueRequestForTLSSecret(c client.Client, logger *slog.Logger) handler.EventHandler {
+	return handler.EnqueueRequestsFromMapFunc(func(ctx context.Context, a client.Object) []reconcile.Request {
+		gateways := helpers.GetGatewaysForSecret(ctx, c, a, logger)
+		reqs := make([]reconcile.Request, 0, len(gateways))
+		for _, gw := range gateways {
+			reqs = append(reqs, reconcile.Request{
+				NamespacedName: types.NamespacedName{
+					Namespace: gw.GetNamespace(),
+					Name:      gw.GetName(),
+				},
+			})
+		}
+		return reqs
+	})
+}
